+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/Asteroid.java
@@ -8,7 +8,7 @@ import java.util.LinkedList;
 public class Asteroid extends SpaceEntity {
 
     public Asteroid(LinkedList<Entity> subscriber, float x, float y) {
-        super(EntityType.ASTEROID, subscriber, EntityTurn.INCOMING, new CantShoot(), EntityShapes.ASTEROID_SHAPE, GameImage.ASTEROID.get(), x, y, 0.02f, 100, 1);
+        super(EntityType.ASTEROID, subscriber, EntityTurn.INCOMING, new CantShoot(), EntityShapes.ASTEROID_SHAPE, GameImage.ASTEROID, x, y, 0.02f, 100, 1);
     }
 
     public Asteroid(float x, float y) {