+++ src/main/java/com/github/mjkuranda/spaceadventure2/resources/GameFont.java
@@ -10,32 +10,30 @@ import java.io.IOException;
 
 public class GameFont implements GameResource<UnicodeFont> {
 
-    public static GameFont PERFECT_DOS_VGA_437;
-    public static GameFont PERFECT_DOS_VGA_437_WIN;
-    public static GameFont REAL_VHS;
-    public static GameFont RETRO_GAMING;
+    public static UnicodeFont PERFECT_DOS_VGA_437;
+    public static UnicodeFont PERFECT_DOS_VGA_437_WIN;
+    public static UnicodeFont REAL_VHS;
+    public static UnicodeFont RETRO_GAMING;
 
     private UnicodeFont uniFont;
 
     public GameFont(String name) throws IOException, FontFormatException, SlickException {
         Font uiFont1 = java.awt.Font.createFont(java.awt.Font.TRUETYPE_FONT,
                 ResourceLoader.getResourceAsStream(getResourcePath("fonts", name)));
-        uiFont1 = uiFont1.deriveFont(Font.PLAIN, 16.f); //You can change "PLAIN" to "BOLD" or "ITALIC"... and 16.f is the size of your font
+        uiFont1 = uiFont1.deriveFont(Font.PLAIN, 16.f);
 
         uniFont = new org.newdawn.slick.UnicodeFont(uiFont1);
         uniFont.addAsciiGlyphs();
-        uniFont.getEffects().add(new ColorEffect(Color.white)); //You can change your color here, but you can also change it in the render{ ... }
+        uniFont.getEffects().add(new ColorEffect(Color.white));
         uniFont.addAsciiGlyphs();
         uniFont.loadGlyphs();
-
-        // PREFIX_PATH + "fonts/yourFontFile.ttf"
     }
 
     public static void load() throws SlickException, IOException, FontFormatException {
-        PERFECT_DOS_VGA_437 = new GameFont("Perfect DOS VGA 437.ttf");
-        PERFECT_DOS_VGA_437_WIN = new GameFont("Perfect DOS VGA 437 Win.ttf");
-        REAL_VHS = new GameFont("RealVhsFontRegular-WyV0z.ttf");
-        RETRO_GAMING = new GameFont("Retro Gaming.ttf");
+        PERFECT_DOS_VGA_437 = new GameFont("Perfect DOS VGA 437.ttf").get();
+        PERFECT_DOS_VGA_437_WIN = new GameFont("Perfect DOS VGA 437 Win.ttf").get();
+        REAL_VHS = new GameFont("RealVhsFontRegular-WyV0z.ttf").get();
+        RETRO_GAMING = new GameFont("Retro Gaming.ttf").get();
     }
 
     @Override